+++ R/taxtable2tree.R
@@ -1,9 +1,0 @@
-#' Taxonomic table to taxonomic tree
-#' @description Compute a taxonomic tree of class phylo from taxonomic table.
-#' @param ... Nothing
-#' @return error
-#' @export
-taxtable2tree <- function(...){
-
-  stop("taxtable2tree() is deprecated, use yatah::taxtree() instead.")
-}